refactor: run rendering tests
